Oss4climate renaming and listing extension (#36)
* [create-pull-request] automated change

* Renaming oss4energy to oss4climate

---------
