+++ repo_index.toml
@@ -713,6 +713,7 @@ repositories = [
     "https://github.com/Open-ET/openet-ssebop",
     "https://github.com/Open-Earth-Foundation/CityCatalyst",
     "https://github.com/Open-Earth-Foundation/OpenClimate",
+    "https://github.com/Open-Environmental-Science/awesome-open-hydrology",
     "https://github.com/Open-MSS/MSS",
     "https://github.com/Open-Power-System-Data/conventional_power_plants",
     "https://github.com/Open-Power-System-Data/renewable_power_plants",
@@ -784,8 +785,9 @@ repositories = [
     "https://github.com/Permian-Global-Research/rsi",
     "https://github.com/PerovskiteSCModelling/IonMonger",
     "https://github.com/PetervanLunteren/EcoAssist",
+    "https://github.com/PhilipMostert/PointedSDMs",
     "https://github.com/Pierre-VF/Enedis-data-io",
-    "https://github.com/Pierre-VF/oss4energy",
+    "https://github.com/Pierre-VF/oss4climate",
     "https://github.com/PierrickP/multicycles",
     "https://github.com/Planheat/Planheat-Tool",
     "https://github.com/PlanktoScope/PlanktoScope",
@@ -803,6 +805,7 @@ repositories = [
     "https://github.com/PowerGridModel/power-grid-model",
     "https://github.com/PredictiveEcology/LandR",
     "https://github.com/Project-AgML/AgML",
+    "https://github.com/ProjectBabbler/ebird-api",
     "https://github.com/ProjectDrawdown/solutions",
     "https://github.com/PublicDomainRelease/GSFLOW",
     "https://github.com/PyPSA/247-cfe",
@@ -855,6 +858,8 @@ repositories = [
     "https://github.com/RichardEssery/FSM2",
     "https://github.com/RobelTakele/AquaBEHER",
     "https://github.com/RoboDonut/awesome-spatial",
+    "https://github.com/Root-Branch/cardamon",
+    "https://github.com/RyanFu008/digitize-heatmap",
     "https://github.com/S2-group/android-runner",
     "https://github.com/S2-group/experiment-runner",
     "https://github.com/S2-group/robot-runner",
@@ -1283,6 +1288,7 @@ repositories = [
     "https://github.com/chrisroadmap/climateforcing",
     "https://github.com/chrisschuerz/SWATrunR",
     "https://github.com/chuhousen/amerifluxr",
+    "https://github.com/ciceroOslo/ciceroscm",
     "https://github.com/circularfashion/cf-circularity-id-standard",
     "https://github.com/cisco-open/green-monitoring",
     "https://github.com/citrineos/citrineos",
@@ -1748,6 +1754,7 @@ repositories = [
     "https://github.com/jrester/tesla_powerwall",
     "https://github.com/jsta/nhdR",
     "https://github.com/jsynowiec/airqmon",
+    "https://github.com/juanb09111/FinnForest",
     "https://github.com/julienmoeys/soiltexture",
     "https://github.com/jupyter-widgets/ipyleaflet",
     "https://github.com/jupyterlab/jupyterlab",
@@ -2218,6 +2225,7 @@ repositories = [
     "https://github.com/rcdaudt/fully_convolutional_change_detection",
     "https://github.com/rcdaudt/patch_based_change_detection",
     "https://github.com/rdbraatz/data-driven-prediction-of-battery-cycle-life-before-capacity-degradation",
+    "https://github.com/re-cinq/aether",
     "https://github.com/reagro/carob",
     "https://github.com/realimpat/Real-Impact-Tracker-Methodology",
     "https://github.com/rebase-energy/enerflow",
@@ -2643,6 +2651,7 @@ projects = [
     "https://gitlab.com/hkex/resipy",
     "https://gitlab.com/hotmaps/potential/WWTP",
     "https://gitlab.com/ibarraespinosa/vein",
+    "https://gitlab.com/lheea/CN-AeroModels",
     "https://gitlab.com/mosaik/mosaik",
     "https://gitlab.com/natural-solutions/reneco/ecoreleve-data",
     "https://gitlab.com/opendatabio/opendatabio",
