+++ src/app/templates/about.html
@@ -7,6 +7,6 @@ This is a proof-of-concept for a search engine for open-source code in the energ
 
 This is not meant as a production ready project at this stage, but more of a proof of concept at this stage. 
 
-Details of the code and indexing can be found <a href="https://github.com/Pierre-VF/oss4energy/">here</a>.
+Details of the code and indexing can be found <a href="https://github.com/Pierre-VF/oss4climate/">here</a>.
 
 {% endblock %}
