+++ src/test/test_imports.py
@@ -1,11 +1,11 @@
 def test_import():
     # The most basic test to at least ensure that all dependencies imported work out
 
-    import oss4energy
-    from oss4energy import cli, scripts
+    import oss4climate
+    from oss4climate import cli, scripts
 
     # Package
-    assert isinstance(oss4energy.__name__, str)
+    assert isinstance(oss4climate.__name__, str)
 
     # Sub-packages
     assert isinstance(cli.__name__, str)
@@ -13,8 +13,8 @@ def test_import():
 
 
 def test_import_src():
-    from oss4energy import src
-    from oss4energy.src import config, database, helpers, log, model, nlp, parsers
+    from oss4climate import src
+    from oss4climate.src import config, database, helpers, log, model, nlp, parsers
 
     assert isinstance(src.__name__, str)
